requirements.txt
tensorflow==1.7
scipy==1.2.1
numpy==1.16.2
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil